% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etcND.R
\name{etc.genND}
\alias{etc.genND}
\title{Generate the Null Distribution of the Threshold Classifier.}
\usage{
etc.genND(n0, n1, c0, c1, pi0)
}
\arguments{
\item{n0}{an integer indicating the number of negative instances in the sample.}

\item{n1}{an integer indicating the number of positive instances in the sample.}

\item{c0}{the cost of misclassifying a negative instance.}

\item{c1}{the cost of misclassifying a positive instance.}

\item{pi0}{a real number between 0 and 1 indicating the percentage of negative instances in the population.}
}
\value{
a list containing three components:
\item{val}{ a vector with the number of possible values than etc can take.}
\item{pos.perm}{ the number of possible permutations for the given number of positives and negatives.}
\item{fav.perm}{ a vector with the number of favorable permutations for every value in val.}
}
\description{
The function offers an algorithmic aproach to generating the null distribution of the etc classifier under the null hypothesis that the distributions of the positive and the negative class are identical.
}
\examples{
etc.genND(25, 27, 1, 3, 0.5)

}

